Fixed Bug:  steamcmd-anywhere was getting location of symlink, not location of script
Running steamcmd directly worked fine.  It pulled the location of the script
file, pushed to that directory, ran steamcmd, and afterwards popped out, as
intended.  When I used a symbolic link to rename the script and move it into the
path (ln -sT <script-path> /usr/bin/steamcmd), I got broken behavior.  The script
now believed it was in /usr/bin instead of /script/path.

For now, I've set the location to the steamcmd directory by hand.
If anyone else needs this environment variable, I'll move it out into
the general environment variable documentation.