Bug Fix: snapshot_droplet failed to consider pre-existing snapshots
It would have been possible to specify a snapshot name that was
already associated with the account.  That's an event that should have
its behavior carefully defined, but snapshot_droplet didn't have the
information at hand to be able to figure out whether or not that was
the case.

We had to pass manager through tearDown and snapshot_droplet to give
it the ability to check at all.  We also changed the behavior of
get_image_by_name so that, when no such image exists, it returns None
rather than raising an exception.