Stub:  Added object for capturing server state saves.
Short names, brief description, longer description, path to the
file(s) representing the savegame.  Also a set of states that the game
server could be in, like initial world (like what we're using for
factorio speedurns), ongoing play, like an ongoing run that we've
decided to stash for the evening, and archived, where we've saved a
game for showing off something , and we mean to never overwrite it.