New Feature: Fleshing out commands for droplet manipulation
About half done, still need to do loading, and listing the running
droplet's uptime.  Also, none of these commands have any sort of
functionality yet, just documentation for how to get the bot to issue
them, and what it is the bot is meant to do once it receives one.