Refactor:  Change the way commands are dispatched
Commands now have a do field, which is supposed to be the code
executed when the command is issued.  All of the actions are defined
as part of MrDo, since it contains the configuration and the droplet
images, the channels for response, and so on.

This probably isn't the final form, but this is way better than the
poor-man's switch we were previously implementing in handle_msg